+++ migrations/020_mission_cost_attributions.down.sql
@@ -1,0 +1,1 @@
+DROP TABLE IF EXISTS mission_cost_attributions;